+++ NEWS.md
@@ -4,6 +4,7 @@ New Features in R6DS 1.1.1
 ==========================
 
 -   Now every classes in R6DS has the common method `toList`.
+-   The `insert` function of the `RBST` class can take more arguments.
 
 New Features in R6DS 1.1.0
 ==========================